+++ custom_model_runner/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - error server is started in case of imports/pipeline failures
 - drum_autofit() helper for sklearn estimator
+- r_lang fit component, pipeline, and template
 ### Changed
 - updated Java predictor dependencies